chore: move dashboard token checker in protocol impl
