+++ core/protocols/dashboard/index.ts
@@ -1,2 +1,3 @@
 export * from "./msg-api.js";
 export * from "./dashboard-api.js";
+export * from "./token.js";
