+++ core/protocols/dashboard/package.json
@@ -41,9 +41,11 @@
     "@fireproof/core-device-id": "workspace:0.0.0",
     "@fireproof/core-runtime": "workspace:0.0.0",
     "@fireproof/core-types-base": "workspace:0.0.0",
+    "@fireproof/core-types-device-id": "workspace:0.0.0",
     "@fireproof/core-types-protocols-cloud": "workspace:0.0.0",
     "@fireproof/core-types-protocols-dashboard": "workspace:0.0.0",
     "@fireproof/vendor": "workspace:0.0.0",
+    "jose": "^6.1.3",
     "zod": "^4.3.5"
   }
 }
