+++ core/types/protocols/dashboard/index.ts
@@ -2,3 +2,5 @@ export * from "./dash-types.js";
 export * from "./fp-api-interface.js";
 export * from "./msg-is.js";
 export * from "./msg-types.js";
+
+export * from "./token.js";
