+++ dashboard/backend/create-handler.ts
@@ -5,7 +5,6 @@ import {
   Lazy,
   LoggerImpl,
   Result,
-  exception2Result,
   param,
   Option,
   Evento,
@@ -18,16 +17,14 @@ import {
   EventoResult,
   EventoSendProvider,
 } from "@adviser/cement";
-import { FPClerkClaim, FPClerkClaimSchema, FPDeviceIDSessionSchema, SuperThis, SuperThisOpts } from "@fireproof/core-types-base";
+import { SuperThisOpts } from "@fireproof/core-types-base";
 import { createFPApiSQLCtx } from "./api.js";
 import type { Env } from "./cf-serve.js";
-import { ensureSuperThis, coerceInt, sts } from "@fireproof/core-runtime";
+import { ensureSuperThis, coerceInt } from "@fireproof/core-runtime";
 import { BaseSQLiteDatabase } from "drizzle-orm/sqlite-core";
 import { ResultSet } from "@libsql/client";
 import { getCloudPubkeyFromEnv } from "./get-cloud-pubkey-from-env.js";
-import { DeviceIdCA, DeviceIdVerifyMsg } from "@fireproof/core-device-id";
-import { jwtVerify } from "jose";
-import { ClerkVerifiedAuth, FPApiSQLCtx, FPApiToken, VerifiedClaimsResult } from "./types.js";
+import { FPApiSQLCtx } from "./types.js";
 import { deleteTenantItem } from "./public/delete-tenant.js";
 import { updateTenantItem } from "./public/update-tenant.js";
 import { createTenantItem } from "./public/create-tenant.js";
@@ -47,8 +44,8 @@ import { getTokenByResultIdItem } from "./public/get-token-by-result-id.js";
 import { getCertFromCsrItem } from "./public/get-cert-from-csr.js";
 import { ensureUserItem } from "./public/ensure-user.js";
 import { ensureCloudTokenItem } from "./public/ensure-cloud-token.js";
-import { VerifyWithCertificateOptions } from "@fireproof/core-types-device-id";
 import { FPApiParameters } from "@fireproof/core-types-protocols-dashboard";
+import { deviceIdCAFromEnv, tokenApi } from "@fireproof/core-protocols-dashboard";
 
 const defaultHttpHeaders = Lazy(() =>
   HttpHeader.from({
@@ -65,173 +62,7 @@ export function DefaultHttpHeaders(...h: CoercedHeadersInit[]): HeadersInit {
     .AsHeaderInit();
 }
 
-class ClerkApiToken implements FPApiToken {
-  readonly sthis: SuperThis;
-  constructor(sthis: SuperThis) {
-    this.sthis = sthis;
-  }
-
-  readonly keysAndUrls = Lazy((): Result<{ keys: string[]; urls: string[] }> => {
-    const keys: string[] = [];
-    const urls: string[] = [];
-    // eslint-disable-next-line no-constant-condition
-    for (let idx = 0; true; idx++) {
-      const suffix = !idx ? "" : `_${idx}`;
-      const key = `CLERK_PUB_JWT_KEY${suffix}`;
-      const url = `CLERK_PUB_JWT_URL${suffix}`;
-      const rEnvVal = this.sthis.env.gets({
-        [key]: param.OPTIONAL,
-        [url]: param.OPTIONAL,
-      });
-      if (rEnvVal.isErr()) {
-        return Result.Err(rEnvVal.Err());
-      }
-      const { [key]: keyVal, [url]: urlVal } = rEnvVal.Ok();
-      if (!keyVal && !urlVal) {
-        // end loop of CLERK_PUB_JWT_KEYn and CLERK_PUB_JWT_URLn
-        break;
-      }
-      if (keyVal) {
-        keys.push(keyVal);
-      }
-      if (urlVal) {
-        urls.push(
-          ...urlVal
-            .split(",")
-            .map((u) => u.trim())
-            .filter((u) => u),
-        );
-      }
-    }
-    return Result.Ok({ keys, urls });
-  });
-
-  async verify(token: string): Promise<Result<VerifiedClaimsResult>> {
-    const { keys, urls } = this.keysAndUrls().Ok();
-
-    const rt = await sts.verifyToken(token, keys, urls, {
-      parseSchema: (payload: unknown): Result<FPClerkClaim> => {
-        const r = FPClerkClaimSchema.safeParse(payload);
-        if (r.success) {
-          return Result.Ok(r.data);
-        } else {
-          // eslint-disable-next-line no-console
-          console.log("FPClerkClaimSchema parse error", payload, r.error);
-          return Result.Err(r.error);
-        }
-      },
-      verifyToken: async (token, key) => {
-        const rPublicKey = await sts.importJWK(key, "RS256");
-        if (rPublicKey.isErr()) {
-          return Result.Err(rPublicKey);
-        }
-        // const pem = await exportSPKI(rPublicKey.Ok().key);
-        // console.log("ClerkApiToken-verify", pem);
-
-        const r = await exception2Result(
-          () => jwtVerify(token, rPublicKey.Ok().key),
-          // ClerkVerifyToken(token, {
-          //   jwtKey: pem,
-          //   // authorizedParties: ["http://localhost:7370"],
-          // }),
-        );
-        // console.log("ClerkApiToken-verify-jwtVerify", r);
-        if (r.isErr()) {
-          return Result.Err(r);
-        }
-        if (!r.Ok()) {
-          return Result.Err("ClerkVerifyToken: failed");
-        }
-        return Result.Ok({
-          payload: r.Ok(),
-        });
-      },
-    });
-    if (rt.isErr()) {
-      return Result.Err(rt.Err());
-    }
-    const t = rt.Ok();
-    return Result.Ok({
-      type: "clerk",
-      token,
-      claims: t.payload,
-    });
-  }
-}
-
-const rDeviceIdCA = Lazy((sthis: SuperThis) => {
-  const rEnv = sthis.env.gets({
-    DEVICE_ID_CA_PRIV_KEY: param.REQUIRED,
-    DEVICE_ID_CA_CERT: param.REQUIRED,
-  });
-  if (rEnv.isErr()) {
-    throw rEnv.Err();
-  }
-  // console.log("rDeviceIdCA Env:", stripper(/^(?!DEVICE_ID)/, rEnv.Ok()));
-  const envVals = rEnv.Ok();
-  return DeviceIdCA.from(
-    sthis,
-    {
-      privateKey: envVals.DEVICE_ID_CA_PRIV_KEY,
-      signedCert: envVals.DEVICE_ID_CA_CERT,
-    },
-    {
-      generateSerialNumber: async () => sthis.nextId(32).str,
-    },
-  );
-});
-
-class DeviceIdApiToken implements FPApiToken {
-  readonly sthis: SuperThis;
-  readonly opts: VerifyWithCertificateOptions;
-  constructor(sthis: SuperThis, opts: VerifyWithCertificateOptions) {
-    this.sthis = sthis;
-    this.opts = opts;
-  }
-  async verify(token: string): Promise<Result<VerifiedClaimsResult>> {
-    const rCA = await rDeviceIdCA(this.sthis);
-    if (rCA.isErr()) {
-      return Result.Err(rCA.Err());
-    }
-    const verify = new DeviceIdVerifyMsg(this.sthis.txt.base64, [(await rCA.Ok().caCertificate()).Ok()], {
-      maxAge: 3600,
-      ...this.opts,
-    });
-    const res = await verify.verifyWithCertificate(token, FPDeviceIDSessionSchema);
-    if (res.valid) {
-      const creatingUser = (res.certificate.certificate.asCert() as unknown as { creatingUser: ClerkVerifiedAuth }).creatingUser;
-      // console.log("DeviceIdApiToken-verify", Object.keys(res.certificate.certificate.asCert()))
-      // console.log("DeviceIdApiToken-verify", creatingUser);
-      if (!creatingUser || creatingUser.type !== "clerk") {
-        return Result.Err(`DeviceIdApiToken-verify: unsupported creatingUser type: ${creatingUser}`);
-      }
-      // console.log("DeviceIdApiToken-verify-1", JSON.stringify(creatingUser.claims));
-      return Result.Ok({
-        type: "device-id",
-        token,
-        claims: creatingUser.claims,
-      });
-    }
-    return Result.Err(res.error);
-  }
-}
-
 export type DashSqlite = BaseSQLiteDatabase<"async", ResultSet | D1Result, Record<string, never>>;
-
-const tokenApi = Lazy(async (sthis: SuperThis) => {
-  // const rDeviceIdCA = await DeviceIdCA.from(sthis, {
-  //   privateKeyEnv: "DEVICE_ID_CA_PRIV_KEY",
-  //   signedCertEnv: "DEVICE_ID_CA_CERT",
-  // }, {
-  //   generateSerialNumber: async () => sthis.nextId(32).str,
-  // });
-  return {
-    "device-id": new DeviceIdApiToken(sthis, {
-      clockTolerance: 60,
-    }),
-    clerk: new ClerkApiToken(sthis),
-  };
-});
 
 export type BindPromise<T> = (promise: Promise<T>) => Promise<T>;
 
@@ -460,18 +291,7 @@ export async function createHandler<T extends DashSqlite>(db: T, env: Record<str
     throw rCloudPublicKey.Err();
   }
 
-  // Create DeviceIdCA from environment variables
-  const rDeviceIdCA = await DeviceIdCA.from(
-    sthis,
-    {
-      privateKey: envVals.DEVICE_ID_CA_PRIV_KEY,
-      signedCert: envVals.DEVICE_ID_CA_CERT,
-    },
-    {
-      generateSerialNumber: async () => sthis.nextId(32).str,
-    },
-  );
-
+  const rDeviceIdCA = await deviceIdCAFromEnv(sthis);
   if (rDeviceIdCA.isErr()) {
     throw rDeviceIdCA.Err();
   }
@@ -511,9 +331,19 @@ export async function createHandler<T extends DashSqlite>(db: T, env: Record<str
         return;
     }
   });
+
   const fpApiCtx = new AppContext().set(
     "fpApiCtx",
-    createFPApiSQLCtx(sthis, db, await tokenApi(sthis), rDeviceIdCA.Ok(), svcParams),
+    createFPApiSQLCtx(
+      sthis,
+      db,
+      await tokenApi(sthis, {
+        clockTolerance: 60,
+        deviceIdCA: rDeviceIdCA.Ok(),
+      }),
+      rDeviceIdCA.Ok(),
+      svcParams,
+    ),
   );
   const evento = fpApiEvento();
   const send = new SendResponseProvider();
